second commit - step 2

+++ src/main/java/com/example/bookInventory/respository/BookConditionRepository.java
@@ -1,10 +1,12 @@
 package com.example.bookInventory.respository;
 
+
 import com.example.bookInventory.entity.BookCondition;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookConditionRepository extends JpaRepository<BookCondition, Long> {
+public interface BookConditionRepository extends JpaRepository<BookCondition, Integer> {
     // You can add custom query methods here if needed
 }
+
